requirements.txt
flox-lib
python-dotenv
typing
youdaoai